This is a csh script that generates an imaging protocol file for use by the ITC18. This script is complicated and shouldn't be used as is.
Camera trigger is assumed to be connect to TTL output 0.
Shutters 1,2,3 are connected to TTL ouput 1,2,3.
EPI and TIRF are switches that activate shutters 4,5 connected to TTL output 4 and 5.
Filter positions are for a Sutter 10-2, using TTL outputs 8-15 found on the back of the ITC18.

This script is for a very advanced user.