Remove pymmcore DCam example
The correct way to load the demo camera is to load DHub and DCam, but doing this correctly is an advanced topic and probably not appropriate for this quick introduction. Using a config file is the recommended method for all but relatively advanced developers. It would be good to somewhere document the correct way to load devices manually, but it does not belong here, so removing the example.